Added a new ChainConfig property and tests
A new property `empty_consensus_reward_maturity_distance` has been added to
the ChainConfig so that it is possible to specify a non-zero block reward
maturity even when the consensus is set to `None`. This is used for testing.

Added tests for 100% coverage of the
`ChainstateRef::check_block_reward_maturity_settings` failure cases.